The spaceship can now shoot projectiles

+++ Gradius/src/com/gradiuss/game/models/Projectile.java
@@ -4,7 +4,7 @@ import android.graphics.Bitmap;
 
 public abstract class Projectile extends MovingObject {
 
-	public Projectile(Bitmap bitmap, int x, int y) {
+	public Projectile(Bitmap bitmap, float x, float y) {
 		super(bitmap, x, y);
 	}
